refactor(sidecar): the scene reader leaves the handler it was living in
terra/actions.py held the Sentinel-2 reader as well as the seventeen actions
that call it: scene discovery over two sources, band clipping, the digital
number to reflectance conversion, the reference grid, and the vegetation
indices. None of that is an action, and none of it is specific to one product.
It is now terra/imagery/, and actions.py drops from 4,271 lines to 3,898.

sentinel2.py takes the reader. The two conversions stay two functions, because
the seam between them is deliberate and a single one with a flag is a seam
somebody closes by accident: to_reflectance is the physically correct
conversion and feeds everything the application reports as a quantity, while
as_trained omits the baseline 04.00 offset because the shipped heads were
fitted without it. The reasoning for that, measured over a Cascavel AOI,
travels with the function that carries it.

indices.py takes NDVI, EVI and SAVI. compute_index_features stays behind: the
fourteen temporal descriptors it builds are the trained model's feature
convention, not a property of a scene, and it moves with the land cover slice.

cog.py takes the GDAL settings every remote read needs applied first, as
cog.configure(). It is not Sentinel-2 specific -- the DEM reads through it too
-- so it is its own module rather than a function inside the scene reader.

Also removed: three imports nothing used. Two were left behind by this cut,
and one was a rasterio reproject aliased inside action_solar_terrain and never
called. With no linter wired up, nothing said so; that is an argument for
wiring one up, made here in the smallest possible way.

Verification: pytest 464 passed, 8 skipped. gofmt, go vet and go test -race
clean on internal/analysis and the root package. A ping through the entry point
answers, and resolving an action still imports none of pvlib, torch, sklearn or
pystac_client.
