+++ sidecar/terra/imagery/__init__.py
@@ -1,0 +1,7 @@
+"""
+What a scene is read as: the products of a catalogue, their bands, and the
+quantities computed from them before any product decides what they mean.
+
+Nothing here knows about land cover, water or energy. A module that needs the
+legend of a classification belongs with that classification, not here.
+"""
